Changed unknown floats to uints
It's much safer to assume that unknown values are integers

+++ HaloOnlineTagTool/Resources/Geometry/GeometryReference.cs
@@ -28,15 +28,15 @@ namespace HaloOnlineTagTool.Resources.Geometry
 
 		public List<UnknownNodeyBlock> UnknownNodey;
 		public List<UnknownBlock> Unknown2;
-		public float Unknown3;
-		public float Unknown4;
-		public float Unknown5;
+		public uint Unknown3;
+		public uint Unknown4;
+		public uint Unknown5;
 		public List<UnknownSection> UnknownSections;
 		public List<NodeMap> NodeMaps;
 		public List<UnknownBlock2> Unknown6;
-		public float Unknown7;
-		public float Unknown8;
-		public float Unknown9;
+		public uint Unknown7;
+		public uint Unknown8;
+		public uint Unknown9;
 		public List<UnknownYoBlock> UnknownYo;
 
 		/// <summary>
@@ -49,21 +49,21 @@ namespace HaloOnlineTagTool.Resources.Geometry
 		[TagStructure(Size = 0x30)]
 		public class UnknownNodeyBlock
 		{
-			public float Unknown;
-			public float Unknown2;
-			public float Unknown3;
-			public float Unknown4;
-			public float Unknown5;
-			public float Unknown6;
-			public float Unknown7;
-			public float Unknown8;
+			public uint Unknown;
+			public uint Unknown2;
+			public uint Unknown3;
+			public uint Unknown4;
+			public uint Unknown5;
+			public uint Unknown6;
+			public uint Unknown7;
+			public uint Unknown8;
 			public sbyte NodeIndex;
 			public sbyte NodeIndex2;
 			public sbyte NodeIndex3;
 			public sbyte NodeIndex4;
-			public float Unknown9;
-			public float Unknown10;
-			public float Unknown11;
+			public uint Unknown9;
+			public uint Unknown10;
+			public uint Unknown11;
 		}
 
 		[TagStructure(Size = 0x18)]
@@ -95,27 +95,27 @@ namespace HaloOnlineTagTool.Resources.Geometry
 			[TagStructure(Size = 0x30)]
 			public class UnknownBlock
 			{
-				public float Unknown;
-				public float Unknown2;
-				public float Unknown3;
-				public float Unknown4;
-				public float Unknown5;
-				public float Unknown6;
-				public float Unknown7;
-				public float Unknown8;
-				public float Unknown9;
-				public float Unknown10;
-				public float Unknown11;
-				public float Unknown12;
+				public uint Unknown;
+				public uint Unknown2;
+				public uint Unknown3;
+				public uint Unknown4;
+				public uint Unknown5;
+				public uint Unknown6;
+				public uint Unknown7;
+				public uint Unknown8;
+				public uint Unknown9;
+				public uint Unknown10;
+				public uint Unknown11;
+				public uint Unknown12;
 			}
 		}
 
 		[TagStructure(Size = 0x10)]
 		public class UnknownYoBlock
 		{
-			public float Unknown;
-			public float Unknown2;
-			public float Unknown3;
+			public uint Unknown;
+			public uint Unknown2;
+			public uint Unknown3;
 			public short UnknownIndex;
 			public short Unknown4;
 		}
