+++ HaloOnlineTagTool/TagStructures/AiGlobals.cs
@@ -13,14 +13,14 @@ namespace HaloOnlineTagTool.TagStructures
 	public class AiGlobals
 	{
 		public List<UnknownBlock> Unknown;
-		public float Unknown2;
+		public uint Unknown2;
 
 		[TagStructure(Size = 0x144)]
 		public class UnknownBlock
 		{
-			public float Unknown;
-			public float Unknown2;
-			public float Unknown3;
+			public uint Unknown;
+			public uint Unknown2;
+			public uint Unknown3;
 			public float DangerBroadlyFacing;
 			public float DangerShootingNear;
 			public float DangerShootingAt;
@@ -53,41 +53,41 @@ namespace HaloOnlineTagTool.TagStructures
 			public float VaultStepMax;
 			public float VaultCrouchMin;
 			public float VaultCrouchMax;
-			public float Unknown4;
-			public float Unknown5;
-			public float Unknown6;
-			public float Unknown7;
+			public uint Unknown4;
+			public uint Unknown5;
+			public uint Unknown6;
+			public uint Unknown7;
 			public List<GravemindProperty> GravemindProperties;
 			public float ScaryTargetThreshold;
 			public float ScaryWeaponThreshold;
-			public float Unknown8;
-			public float Unknown9;
-			public float Unknown10;
-			public float Unknown11;
-			public float Unknown12;
-			public float Unknown13;
-			public float Unknown14;
-			public float Unknown15;
-			public float Unknown16;
-			public float Unknown17;
-			public float Unknown18;
+			public uint Unknown8;
+			public uint Unknown9;
+			public uint Unknown10;
+			public uint Unknown11;
+			public uint Unknown12;
+			public uint Unknown13;
+			public uint Unknown14;
+			public uint Unknown15;
+			public uint Unknown16;
+			public uint Unknown17;
+			public uint Unknown18;
 			public List<Style> Styles;
 			public List<Formation> Formations;
 			public List<SquadTemplate> SquadTemplates;
-			public float Unknown19;
-			public float Unknown20;
-			public float Unknown21;
-			public float Unknown22;
-			public float Unknown23;
-			public float Unknown24;
-			public float Unknown25;
-			public float Unknown26;
-			public float Unknown27;
-			public float Unknown28;
-			public float Unknown29;
-			public float Unknown30;
-			public float Unknown31;
-			public float Unknown32;
+			public uint Unknown19;
+			public uint Unknown20;
+			public uint Unknown21;
+			public uint Unknown22;
+			public uint Unknown23;
+			public uint Unknown24;
+			public uint Unknown25;
+			public uint Unknown26;
+			public uint Unknown27;
+			public uint Unknown28;
+			public uint Unknown29;
+			public uint Unknown30;
+			public uint Unknown31;
+			public uint Unknown32;
 
 			[TagStructure(Size = 0xC)]
 			public class GravemindProperty
