+++ HaloOnlineTagTool/TagStructures/Antenna.cs
@@ -15,15 +15,15 @@ namespace HaloOnlineTagTool.TagStructures
 		public StringId AttachmentMarkerName;
 		public HaloTag Bitmaps;
 		public HaloTag Physics;
-		public float Unknown;
-		public float Unknown2;
-		public float Unknown3;
-		public float Unknown4;
-		public float Unknown5;
-		public float Unknown6;
-		public float Unknown7;
+		public uint Unknown;
+		public uint Unknown2;
+		public uint Unknown3;
+		public uint Unknown4;
+		public uint Unknown5;
+		public uint Unknown6;
+		public uint Unknown7;
 		public List<Vertex> Vertices;
-		public float Unknown8;
+		public uint Unknown8;
 
 		[TagStructure(Size = 0x40)]
 		public class Vertex
@@ -42,9 +42,9 @@ namespace HaloOnlineTagTool.TagStructures
 			public float LodColorG;
 			public float LodColorB;
 			public float Width;
-			public float Unknown2;
-			public float Unknown3;
-			public float Unknown4;
+			public uint Unknown2;
+			public uint Unknown3;
+			public uint Unknown4;
 		}
 	}
 }
